Allow access to PrivateAssetsBucket via instance profile, and consolidate common IAM policies
