+++ stack/logs.py
@@ -1,5 +1,6 @@
-from troposphere import logs
+from troposphere import iam, Join, logs
 
+from .common import arn_prefix
 from .template import template
 
 
@@ -8,4 +9,22 @@ container_log_group = logs.LogGroup(
     template=template,
     RetentionInDays=365,
     DeletionPolicy="Retain",
+)
+
+
+logging_policy = iam.Policy(
+    PolicyName="LoggingPolicy",
+    PolicyDocument=dict(
+        Statement=[dict(
+            Effect="Allow",
+            Action=[
+                "logs:Create*",
+                "logs:PutLogEvents",
+            ],
+            Resource=Join("", [
+                arn_prefix,
+                ":logs:*:*:*",  # allow logging to any log group
+            ]),
+        )],
+    ),
 )
